// example4/oneteam/src/views/BaseAccess/VendorForm/VendorFormSections/VendorFormSections.types.ts
export type commonInfoFields =
  | 'orgName'
  | 'inn'
  | 'kpp'
  | 'orgCategory'
  | 'goodsCategory'
  | 'goodsExtraCategory'
  | 'brands'
  | 'orgSite'
  | 'orgCatalogLink';
export type contactInfoFields =
  | 'fio'
  | 'position'
  | 'email'
  | 'phone'
  | 'deJuroAddress'
  | 'deFactoAddress'
  | 'warehouses'
  | 'holdings';
export type organizationInfoFields = 'orgPartners' | 'sum' | 'files';
export type contractPreviewFields = 'signTheContract' | 'typeOfContract';